Moved authentication logic from MainActivity to AuthManager
