+++ app/src/main/java/com/zybooks/workway/MainActivity.kt
@@ -2,37 +2,16 @@ package com.zybooks.workway
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.credentials.ClearCredentialStateRequest
-import androidx.credentials.CredentialManager
-import androidx.credentials.CustomCredential
-import androidx.credentials.GetCredentialRequest
-import androidx.credentials.exceptions.ClearCredentialException
-import androidx.credentials.exceptions.GetCredentialException
-import androidx.lifecycle.lifecycleScope
-import com.google.android.libraries.identity.googleid.GetGoogleIdOption
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential.Companion.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL
-import com.google.firebase.Firebase
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.GoogleAuthProvider
-import com.google.firebase.auth.auth
-import com.zybooks.workway.model.User
-import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var auth: FirebaseAuth
-    private lateinit var credentialManager: CredentialManager
+    private lateinit var authManager: AuthManager
     private lateinit var activityResultLauncher: ActivityResultLauncher<Intent>
-
-    private var loggingIn = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -44,102 +23,22 @@ class MainActivity : AppCompatActivity() {
             insets
         }
 
-        auth = Firebase.auth
-        credentialManager = CredentialManager.create(baseContext)
         activityResultLauncher = registerForActivityResult(
             ActivityResultContracts.StartActivityForResult()
         ) { result ->
             if (result.resultCode == RESULT_OK) {
                 val data = result.data?.getBooleanExtra("LOGOUT", false)
-                if (data == true) logOut()
+                if (data == true) authManager.logOut()
             }
         }
+        authManager = AuthManager(this, activityResultLauncher)
     }
 
     override fun onResume() {
         super.onResume()
-        if (loggingIn) return
-        if (auth.currentUser != null) activityResultLauncher.launch(Intent(this, HomeScreen::class.java))
-        else launchCredentialManager()
-    }
-
-    private fun launchCredentialManager() {
-        val googleIdOption = GetGoogleIdOption.Builder()
-            .setServerClientId(getString(R.string.web_client_id))
-            .setFilterByAuthorizedAccounts(false)
-            .build()
-
-        val request = GetCredentialRequest.Builder()
-            .addCredentialOption(googleIdOption)
-            .build()
-
-        lifecycleScope.launch {
-            try {
-                val result = credentialManager.getCredential(
-                    context = baseContext,
-                    request = request
-                )
-                handleSignIn(result.credential)
-            } catch (e: GetCredentialException) {
-                Log.e(TAG, "Couldn't retrieve user's credentials: ${e.localizedMessage}")
-                launchCredentialManager()
-            }
-        }
-    }
-
-    private fun handleSignIn(credential: androidx.credentials.Credential) {
-        if (credential is CustomCredential && credential.type == TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
-            val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
-            firebaseAuthWithGoogle(googleIdTokenCredential.idToken)
-        } else {
-            Log.w(TAG, "Credential is not of type Google ID!")
-        }
-    }
-
-    private fun firebaseAuthWithGoogle(idToken: String) {
-        loggingIn = true
-        val credential = GoogleAuthProvider.getCredential(idToken, null)
-        auth.signInWithCredential(credential)
-            .addOnCompleteListener(this) { task ->
-                if (task.isSuccessful) {
-                    logIn()
-                } else {
-                    loggingIn = false
-                    Log.w(TAG, "signInWithCredential:failure", task.exception)
-                }
-            }
-    }
-
-    private fun logIn() {
-        val user: User
-        if (false /*getUser(auth.currentUser?.uid) != null*/) { // User exists
-            // Load user attributes from database
-            user = User()
-        } else { // User is new
-            user = User().apply {
-                userID = auth.currentUser?.uid
-                name = auth.currentUser?.displayName
-                email = auth.currentUser?.email
-            }
-        }
-        val text = "Logged in as ${user.email}"
-        Toast.makeText(this, text, Toast.LENGTH_SHORT).show()
-        activityResultLauncher.launch(Intent(this, HomeScreen::class.java))
-    }
-
-    private fun logOut() {
-        auth.signOut()
-        lifecycleScope.launch {
-            try {
-                val clearRequest = ClearCredentialStateRequest()
-                credentialManager.clearCredentialState(clearRequest)
-            } catch (e: ClearCredentialException) {
-                Log.e(TAG, "Couldn't clear user credentials: ${e.localizedMessage}")
-            }
-        }
-    }
-
-    companion object {
-        private const val TAG = "GoogleActivity"
+        if (authManager.loggingIn) return
+        if (authManager.isLoggedIn())
+            activityResultLauncher.launch(Intent(this, HomeScreen::class.java))
+        else authManager.launchCredentialManager()
     }
 }
